% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtm.R
\name{calc_vtms}
\alias{calc_vtms}
\title{Compute variable threshold metrics for classifier performance.}
\usage{
calc_vtms(y_true, y_pred)
}
\arguments{
\item{y_true}{true binary labels in binary label indicators}

\item{y_pred}{target scores, can either be probability estimates of
the positive class, confidence values, or binary decisions.}
}
\description{
Compute roc and prc curves as well as AUCs (area under the curves).
}

