% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtm.R
\name{get_tjur}
\alias{get_tjur}
\title{Compute the coefficient of discrimination.}
\usage{
get_tjur(y_true, y_pred)
}
\arguments{
\item{y_true}{true binary labels in binary label indicators}

\item{y_pred}{target scores, can either be probability estimates of
the positive class, confidence values, or binary decisions.}
}
\description{
Compute Tjur's R-squared. See https://doi.org/10.1198/tast.2009.08210
}

