% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet_coefs}
\alias{glmnet_coefs}
\title{Extract glmnet model coefficients.}
\usage{
glmnet_coefs(cv_glmnet, X, y, s = "lambda.1se", prepend = "", ...)
}
\arguments{
\item{cv_glmnet}{a \code{\link[glmnet]{cv.glmnet}} object to extract
coefficients for.}

\item{X}{matrix of features used to train \code{cv_glmnet}}

\item{y}{outcome used to train \code{cv_glmnet}}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required}

\item{prepend}{text to prepend to the \code{coef} and \code{zcoef} column names}
}
\description{
Extract standardized and unstandardized coefficients.
}

