% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet_predict}
\alias{glmnet_predict}
\title{Predictions from a glmnet object}
\usage{
glmnet_predict(cv_glmnet, X, s = "lambda.1se")
}
\arguments{
\item{cv_glmnet}{a \code{\link[glmnet]{cv.glmnet}} object.}

\item{X}{matrix of features to predict from.}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required.}
}
\description{
Predictions from a glmnet object
}

