% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet_train}
\alias{glmnet_train}
\title{Train a regularized logistic regression model.}
\usage{
glmnet_train(X, y, w, alpha = 1, s = "lambda.1se", cores = 7, seed = 0,
  ...)
}
\description{
Train a regularized logistic regression model.
}

