% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtm.R
\name{prune_prc}
\alias{prune_prc}
\title{Prune points from a precision-recall curve.}
\usage{
prune_prc(prc_df, min_distance = 5e-04)
}
\arguments{
\item{prc_df}{a dataframe containing a precision-recall curve
created by \code{\link{calc_vtms}}.}

\item{min_distance}{minimum distance between two retained points}
}
\description{
Reduces the number of points in a PRC when many observations.
}

