% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtm.R
\name{prune_roc}
\alias{prune_roc}
\title{Prune points from a ROC curve.}
\usage{
prune_roc(roc_df)
}
\arguments{
\item{roc_df}{a dataframe containing a ROC curve
created by \code{\link{calc_vtms}}.}
}
\description{
Reduces the number of points in an ROC curve when many observations
}

