% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_dhimmel}
\alias{theme_dhimmel}
\title{The official ggplot2 theme of dhimmel.}
\usage{
theme_dhimmel(base_size = 11.5, base_family = "")
}
\description{
The theme is a modified version of ggplot2::theme_bw to improve space
conservation and improve color aesthetics.
}

